+++ packages/auditable-item-stream-service/src/entities/auditableItemStreamEntry.ts
@@ -1,7 +1,7 @@
 // Copyright 2024 IOTA Stiftung.
 // SPDX-License-Identifier: Apache-2.0.
-import { JsonLdTypes, type IJsonLdNodeObject } from "@gtsc/data-json-ld";
-import { entity, property } from "@gtsc/entity";
+import { JsonLdTypes, type IJsonLdNodeObject } from "@twin.org/data-json-ld";
+import { entity, property } from "@twin.org/entity";
 
 /**
  * Class describing the auditable item stream entry.