+++ packages/auditable-item-stream-models/docs/changelog.md
@@ -1,5 +1,5 @@
-# @gtsc/auditable-item-stream-models - Changelog
+# @twin.org/auditable-item-stream-models - Changelog
 
-## v0.0.3
+## v0.0.1-next.3
 
 - Initial Release